rename lambda.py due to lambda being a reserved word
